The wmts bounds calculation could overflow and become negative.
The tilecache bounds calculation in wmts was performed as an integer which is too small to contain all possible, this calculation is now done as a double.

+++ core/src/main/java/org/mapfish/print/map/tiled/wmts/Matrix.java
@@ -36,7 +36,7 @@ public class Matrix {
     /**
      * A 2 dimensional array containing number of tiles in the matrix for the columns (0) and rows (1).
      */
-    public int[] matrixSize;
+    public double[] matrixSize;
     /**
      * The scale of the matrix.
      */
